#68 Bug/Uncaught TypeError: Cannot read properties of null (reading 'hasChildNodes')
