+++ src/assets/css/SearchMap/Toggle/VariablePlace.css
@@ -1,4 +1,4 @@
-#category {
+.category {
   position: absolute;
   margin: 70px 100px;
   border-radius: 5px;
@@ -6,10 +6,10 @@
   box-shadow: 0 1px 1px rgba(0, 0, 0, 0.4);
   background: #fff;
   overflow: hidden;
-  z-index: 2;
+  z-index: 10;
 }
 
-#category #CE7 {
+.category #CE7 {
   list-style: none;
   width: 200px;
   border-right: 1px solid #acacac;
@@ -19,27 +19,21 @@
   font-weight: bold;
 }
 
-#category #CE7:hover {
-  background: #ffe6e6;
-  border-left: 1px solid #acacac;
-  margin-left: -1px;
-}
-
-#category #CE7 span {
+.category #CE7 span {
   display: block;
   margin: 0 auto 3px;
   width: 27px;
   height: 28px;
 }
-#category #CE7 .category_bg {
+.category #CE7 .category_bg {
   background: url(https://t1.daumcdn.net/localimg/localimages/07/mapapidoc/places_category.png)
     no-repeat;
 }
 
-#category #CE7 .cafe {
+.category #CE7 .cafe {
   background-position: -10px -144px;
 }
 
-#category #CE7.on .category_bg {
+.category #CE7.on .category_bg {
   background-position-x: -46px;
 }
